some fix to run API scrapping

+++ interfaces/job_interface.go
@@ -7,6 +7,6 @@ import(
 type JobRepositoryInterface interface{
 	CreateJob(job model.Job) (int, error)
 	FindJobByRequisitionID(requisition_ID int) (bool, error)
-	FindJobsByRequisitionIDs(requisition_IDs []int) (map[int]bool, error)
-	UpdateLastSeen(requisition_ID int) (int, error)
+	FindJobsByRequisitionIDs(requisition_IDs []int64) (map[int64]bool, error)
+	UpdateLastSeen(requisition_ID int64) (int, error)
 }
